+++ e2e-tests/contentful/cypress/integration/gatsby-image.js
@@ -13,11 +13,11 @@ function hasSVGPlaceholder(el) {
     })
 }
 
-function hasJPEGPlaceholder(el) {
+function hasBase64Placeholder(el) {
   el.children(`img`)
     .should(`have.attr`, `src`)
     .and(src => {
-      expect(src).to.match(/^data:image\/jpeg/)
+      expect(src).to.match(/^data:image\/[a-z]+;base64/)
     })
 }
 
@@ -49,21 +49,21 @@ describe(`gatsby-image`, () => {
   beforeEach(() => {
     cy.visit("/gatsby-image").waitForRouteChange()
   })
-  it(`fluid`, testConfig, () => testGatsbyImage(`fluid`, hasJPEGPlaceholder))
-  it(`fixed`, testConfig, () => testGatsbyImage(`fixed`, hasJPEGPlaceholder))
-  it(`webp`, testConfig, () => testGatsbyImage(`webp`, hasJPEGPlaceholder))
+  it(`fluid`, testConfig, () => testGatsbyImage(`fluid`, hasBase64Placeholder))
+  it(`fixed`, testConfig, () => testGatsbyImage(`fixed`, hasBase64Placeholder))
+  it(`webp`, testConfig, () => testGatsbyImage(`webp`, hasBase64Placeholder))
   it(`traced`, testConfig, () => testGatsbyImage(`traced`, hasSVGPlaceholder))
   it(`sqip`, testConfig, () => testGatsbyImage(`sqip`, hasSVGPlaceholder))
 
   it(`english`, testConfig, () => {
-    testGatsbyImage(`english`, hasJPEGPlaceholder)
+    testGatsbyImage(`english`, hasBase64Placeholder)
     cy.get(`[data-cy="english"] p strong`).should(
       "have.text",
       "Locale - American English (png)"
     )
   })
   it(`german`, testConfig, () => {
-    testGatsbyImage(`german`, hasJPEGPlaceholder)
+    testGatsbyImage(`german`, hasBase64Placeholder)
     cy.get(`[data-cy="german"] p strong`).should(
       "have.text",
       "Locale - German (png)"