+++ packages/gatsby-plugin-gatsby-cloud/src/gatsby-node.js
@@ -4,6 +4,7 @@ import makePluginData from "./plugin-data"
 import buildHeadersProgram from "./build-headers-program"
 import copyFunctionsManifest from "./copy-functions-manifest"
 import createRedirects from "./create-redirects"
+import createSiteConfig from "./create-site-config"
 import { readJSON } from "fs-extra"
 import { joinPath } from "gatsby-core-utils"
 import { DEFAULT_OPTIONS, BUILD_HTML_STAGE, BUILD_CSS_STAGE } from "./constants"
@@ -107,7 +108,12 @@ exports.onPostBuild = async ({ store, pathPrefix }, userPluginOptions) => {
   }
 
   await Promise.all([
+<<<<<<< HEAD
     buildHeadersProgram(pluginData, pluginOptions),
+=======
+    buildHeadersProgram(pluginData, pluginOptions, reporter),
+    createSiteConfig(pluginData, pluginOptions),
+>>>>>>> fbfe3f63dec23d279a27b54b4057dd611dce74bb
     createRedirects(pluginData, redirects, rewrites),
     copyFunctionsManifest(pluginData),
   ])