+++ packages/gatsby/src/utils/webpack-utils.ts
@@ -181,21 +181,21 @@ export const createWebpackUtils = (
   const isSSR = stage.includes(`html`)
 
   const jsxRuntimeExists = reactHasJsxRuntime()
-  const makeExternalOnly = (original: RuleFactory) => (
-    options = {}
-  ): RuleSetRule => {
-    const rule = original(options)
-    rule.include = vendorRegex
-    return rule
-  }
+  const makeExternalOnly =
+    (original: RuleFactory) =>
+    (options = {}): RuleSetRule => {
+      const rule = original(options)
+      rule.include = vendorRegex
+      return rule
+    }
 
-  const makeInternalOnly = (original: RuleFactory) => (
-    options = {}
-  ): RuleSetRule => {
-    const rule = original(options)
-    rule.exclude = vendorRegex
-    return rule
-  }
+  const makeInternalOnly =
+    (original: RuleFactory) =>
+    (options = {}): RuleSetRule => {
+      const rule = original(options)
+      rule.exclude = vendorRegex
+      return rule
+    }
 
   const loaders: ILoaderUtils = {
     json: (options = {}) => {
@@ -319,9 +319,11 @@ export const createWebpackUtils = (
             const autoprefixerPlugin = autoprefixer({
               overrideBrowserslist,
               flexbox: `no-2009`,
-              ...(((postCSSPlugins.find(
-                plugin => plugin.postcssPlugin === `autoprefixer`
-              ) as unknown) as autoprefixer.ExportedAPI)?.options ?? {}),
+              ...((
+                postCSSPlugins.find(
+                  plugin => plugin.postcssPlugin === `autoprefixer`
+                ) as unknown as autoprefixer.ExportedAPI
+              )?.options ?? {}),
             })
 
             postCSSPlugins.unshift(autoprefixerPlugin)
@@ -677,27 +679,27 @@ export const createWebpackUtils = (
               plugins: [
                 // potentially destructive plugins removed - see https://github.com/gatsbyjs/gatsby/issues/15629
                 // use correct config format and remove plugins requiring specific params - see https://github.com/gatsbyjs/gatsby/issues/31619
-                `removeUselessDefs`,
+                // List of default plugins and their defaults: https://github.com/svg/svgo#built-in-plugins
+                // Last update 2021-08-17
                 `cleanupAttrs`,
                 `cleanupEnableBackground`,
                 `cleanupIDs`,
-                `cleanupListOfValues`,
+                `cleanupListOfValues`, // Default: disabled
                 `cleanupNumericValues`,
                 `collapseGroups`,
                 `convertColors`,
                 `convertPathData`,
-                `convertStyleToAttrs`,
+                `convertStyleToAttrs`, // Default: disabled
                 `convertTransform`,
                 `inlineStyles`,
                 `mergePaths`,
                 `minifyStyles`,
                 `moveElemsAttrsToGroup`,
                 `moveGroupAttrsToElems`,
-                `prefixIds`,
-                `removeAttrs`,
+                `prefixIds`, // Default: disabled
+                `removeAttrs`, // Default: disabled
                 `removeComments`,
                 `removeDesc`,
-                `removeDimensions`,
                 `removeDoctype`,
                 `removeEditorsNSData`,
                 `removeEmptyAttrs`,
@@ -706,17 +708,18 @@ export const createWebpackUtils = (
                 `removeHiddenElems`,
                 `removeMetadata`,
                 `removeNonInheritableGroupAttrs`,
-                `removeOffCanvasPaths`,
-                `removeRasterImages`,
-                `removeScriptElement`,
-                `removeStyleElement`,
+                `removeOffCanvasPaths`, // Default: disabled
+                `removeRasterImages`, // Default: disabled
+                `removeScriptElement`, // Default: disabled
+                `removeStyleElement`, // Default: disabled
                 `removeTitle`,
                 `removeUnknownsAndDefaults`,
                 `removeUnusedNS`,
+                `removeUselessDefs`,
                 `removeUselessStrokeAndFill`,
                 `removeXMLProcInst`,
-                `reusePaths`,
-                `sortAttrs`,
+                `reusePaths`, // Default: disabled
+                `sortAttrs`, // Default: disabled
               ],
             },
           },
@@ -767,8 +770,9 @@ export const createWebpackUtils = (
   plugins.extractStats = (): GatsbyWebpackStatsExtractor =>
     new GatsbyWebpackStatsExtractor()
 
-  plugins.eslintGraphqlSchemaReload = (): GatsbyWebpackEslintGraphqlSchemaReload =>
-    new GatsbyWebpackEslintGraphqlSchemaReload()
+  plugins.eslintGraphqlSchemaReload =
+    (): GatsbyWebpackEslintGraphqlSchemaReload =>
+      new GatsbyWebpackEslintGraphqlSchemaReload()
 
   plugins.virtualModules = (): GatsbyWebpackVirtualModules =>
     new GatsbyWebpackVirtualModules()