+++ packages/gatsby-plugin-netlify-cms/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [5.13.0-next.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-netlify-cms@5.12.0-next.3...gatsby-plugin-netlify-cms@5.13.0-next.0) (2021-08-12)
+
+**Note:** Version bump only for package gatsby-plugin-netlify-cms
+
+# [5.12.0-next.3](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-netlify-cms@5.12.0-next.2...gatsby-plugin-netlify-cms@5.12.0-next.3) (2021-08-09)
+
+**Note:** Version bump only for package gatsby-plugin-netlify-cms
+
+# [5.12.0-next.2](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-netlify-cms@5.12.0-next.1...gatsby-plugin-netlify-cms@5.12.0-next.2) (2021-08-05)
+
+**Note:** Version bump only for package gatsby-plugin-netlify-cms
+
+# [5.12.0-next.1](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-netlify-cms@5.12.0-next.0...gatsby-plugin-netlify-cms@5.12.0-next.1) (2021-08-05)
+
+### Bug Fixes
+
+- **deps:** update dependency mini-css-extract-plugin to v1.6.2 ([#32586](https://github.com/gatsbyjs/gatsby/issues/32586)) ([6c6d4e5](https://github.com/gatsbyjs/gatsby/commit/6c6d4e501764614ad737c88ba5810b6fbbcbff68))
+- **deps:** update minor and patch dependencies for gatsby-plugin-netlify-cms ([#32600](https://github.com/gatsbyjs/gatsby/issues/32600)) ([b7b67a9](https://github.com/gatsbyjs/gatsby/commit/b7b67a9cd21abcbc33cae4338da420dd6af02aa6))
+
 # [5.12.0-next.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-netlify-cms@5.11.0-next.0...gatsby-plugin-netlify-cms@5.12.0-next.0) (2021-07-29)
 
 **Note:** Version bump only for package gatsby-plugin-netlify-cms