+++ packages/gatsby-remark-autolink-headers/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [4.10.0-next.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-autolink-headers@4.9.0-next.3...gatsby-remark-autolink-headers@4.10.0-next.0) (2021-08-12)
+
+**Note:** Version bump only for package gatsby-remark-autolink-headers
+
+# [4.9.0-next.3](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-autolink-headers@4.9.0-next.2...gatsby-remark-autolink-headers@4.9.0-next.3) (2021-08-09)
+
+**Note:** Version bump only for package gatsby-remark-autolink-headers
+
+# [4.9.0-next.2](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-autolink-headers@4.9.0-next.1...gatsby-remark-autolink-headers@4.9.0-next.2) (2021-08-05)
+
+**Note:** Version bump only for package gatsby-remark-autolink-headers
+
+# [4.9.0-next.1](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-autolink-headers@4.9.0-next.0...gatsby-remark-autolink-headers@4.9.0-next.1) (2021-08-05)
+
+**Note:** Version bump only for package gatsby-remark-autolink-headers
+
 # [4.9.0-next.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-autolink-headers@4.8.0-next.0...gatsby-remark-autolink-headers@4.9.0-next.0) (2021-07-29)
 
 **Note:** Version bump only for package gatsby-remark-autolink-headers