+++ packages/gatsby-source-graphql/CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.13.0-next.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-graphql@3.12.0-next.3...gatsby-source-graphql@3.13.0-next.0) (2021-08-12)
+
+**Note:** Version bump only for package gatsby-source-graphql
+
+# [3.12.0-next.3](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-graphql@3.12.0-next.2...gatsby-source-graphql@3.12.0-next.3) (2021-08-09)
+
+**Note:** Version bump only for package gatsby-source-graphql
+
+# [3.12.0-next.2](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-graphql@3.12.0-next.1...gatsby-source-graphql@3.12.0-next.2) (2021-08-05)
+
+**Note:** Version bump only for package gatsby-source-graphql
+
+# [3.12.0-next.1](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-graphql@3.12.0-next.0...gatsby-source-graphql@3.12.0-next.1) (2021-08-05)
+
+### Bug Fixes
+
+- **deps:** update dependency @graphql-tools/wrap to ^7.0.8 for gatsby-source-graphql ([#32572](https://github.com/gatsbyjs/gatsby/issues/32572)) ([e5d8603](https://github.com/gatsbyjs/gatsby/commit/e5d860380cb2f6a1dc4dc6ee29b2cdaab02ca670))
+
 # [3.12.0-next.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-graphql@3.11.0-next.0...gatsby-source-graphql@3.12.0-next.0) (2021-07-29)
 
 **Note:** Version bump only for package gatsby-source-graphql